requirements.txt
allure-pytest==2.8.40
allure-python-commons==2.8.40
androguard==3.3.5
apipkg==1.5
Appium-Python-Client==1.1.0
asn1crypto==1.4.0
atomicwrites==1.4.0
attrs==20.3.0
backcall==0.2.0
BeautifulReport==0.1.3
cached-property==1.5.2
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
colorama==0.4.4
colored==1.4.2
cycler==0.10.0
ddt==1.4.2
decorator==4.4.2
et-xmlfile==1.0.1
execnet==1.8.0
future==0.18.2
idna==2.10
importlib-metadata==3.9.0
iniconfig==1.1.1
ipython==7.22.0
ipython-genutils==0.2.0
jedi==0.18.0
kiwisolver==1.3.1
lml==0.1.0
logzero==1.7.0
lxml==4.6.3
matplotlib==3.4.0
networkx==2.5
numpy==1.20.2
openpyxl==3.0.7
packaging==20.9
parso==0.8.1
pickleshare==0.7.5
Pillow==8.1.2
pluggy==0.13.1
prompt-toolkit==3.0.18
py==1.10.0
pydot==1.4.2
pyexcel-io==0.6.4
pyexcel-xls==0.6.2
Pygments==2.8.1
pyparsing==2.4.7
pytest==6.2.2
pytest-assume==2.4.2
pytest-dependency==0.5.1
pytest-forked==1.3.0
pytest-html==3.1.1
pytest-metadata==1.11.0
pytest-ordering==0.6
pytest-repeat==0.9.1
pytest-rerunfailures==9.1.1
pytest-xdist==2.2.1
python-dateutil==2.8.1
PyYAML==5.4.1
redis==3.5.3
requests==2.25.1
selenium==3.141.0
simple-tornado==0.2.0
six==1.15.0
tidevice==0.3.5
toml==0.10.2
tornado==6.1
traitlets==5.0.5
typing-extensions==3.7.4.3
urllib3==1.26.4
wcwidth==0.2.5
xlrd==1.2.0
xlwt==1.3.0
zipp==3.4.1
